Split amp-lightbox component files
