+++ extensions/amp-lightbox/1.0/component.jss.js
