+++ extensions/amp-lightbox/1.0/component.type.js
